More refactoring. Fixing factories so that methods are recognised in json files.
