+++ source/core/assets/configs/leaderboard.txt
@@ -41,3 +41,4 @@ DirtyDefault71230:693,
 DirtyDefault51830:935,410,
 DirtyDefault549:602,
 DirtyDefault39550:-2745,
+DirtyDefault65981:
