--- source/core/src/main/com/deco2800/game/areas/home/Doorway.java
@@ -1,5 +1,0 @@
-package com.deco2800.game.areas.home;
-
-public class Doorway {
-
-}
