--- source/core/src/main/com/deco2800/game/areas/home/Home.java
@@ -1,7 +1,0 @@
-package com.deco2800.game.areas.home;
-
-import com.deco2800.game.entities.Entity;
-
-public class Home {
-    public Entity player;
-}
