+++ source/core/src/main/com/deco2800/game/maps/floor/Doorway.java
@@ -1,0 +1,5 @@
+package com.deco2800.game.maps.floor;
+
+public class Doorway {
+
+}
