+++ source/core/src/main/com/deco2800/game/maps/floor/rooms/Living.java
@@ -1,11 +1,8 @@
-package com.deco2800.game.areas.home.rooms;
+package com.deco2800.game.maps.floor.rooms;
 
 import com.badlogic.gdx.math.GridPoint2;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.ObjectMap;
-import com.deco2800.game.areas.home.Room;
-import com.deco2800.game.areas.home.RoomObject;
-import com.deco2800.game.areas.home.RoomProperties;
 
 public class Living extends Room {
