+++ source/core/src/test/com/deco2800/game/maps/terrain/TerrainComponentTest.java
@@ -1,14 +1,9 @@
-package com.deco2800.game.areas.terrain;
+package com.deco2800.game.maps.terrain;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 
-import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.maps.tiled.TiledMap;
-import com.badlogic.gdx.maps.tiled.TiledMapRenderer;
-import com.badlogic.gdx.math.Vector2;
 import com.deco2800.game.extensions.GameExtension;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 @ExtendWith(GameExtension.class)
